Restructure blankx
